functions-python/requirements.txt
azure-functions==1.17.0
azure-storage-blob==12.19.0
azure-search-documents==11.4.0
azure-cognitiveservices-vision-computervision==0.9.0
azure-cognitiveservices-vision-formrecognizer==3.2.0
openai==1.3.0
PyPDF2==3.0.1
python-docx==1.1.0
Pillow==10.1.0
pytesseract==0.3.10
python-multipart==0.0.6
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0 